pyproject.toml
[project]
name = "agent-lang-stream-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "ansible==11.9.0",
    "ansible-core==2.18.8",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "blinker==1.9.0",
    "cachetools==6.1.0",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "cryptography==45.0.6",
    "distro==1.9.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "langchain==0.3.27",
    "langchain-core==0.3.74",
    "langchain-mcp-adapters==0.1.9",
    "langchain-openai==0.3.30",
    "langchain-text-splitters==0.3.9",
    "langgraph==0.6.4",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.6.4",
    "langgraph-sdk==0.2.0",
    "langsmith==0.4.14",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "mcp[cli]==1.12.4",
    "mdurl==0.1.2",
    "narwhals==2.1.1",
    "numpy==2.3.2",
    "openai==1.99.9",
    "orjson==3.11.2",
    "ormsgpack==1.10.0",
    "packaging==25.0",
    "pandas==2.3.1",
    "pillow==11.3.0",
    "protobuf==6.31.1",
    "pyarrow==21.0.0",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "regex==2025.7.34",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "resolvelib==1.0.1",
    "rich==14.1.0",
    "rpds-py==0.27.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "sse-starlette==3.0.2",
    "starlette==0.47.2",
    "streamlit==1.48.0",
    "tenacity==9.1.2",
    "tiktoken==0.11.0",
    "toml==0.10.2",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "typer==0.16.0",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "watchdog==6.0.0",
    "xxhash==3.5.0",
    "zstandard==0.23.0",
]
